Try to fix multi-arch images
For some reason, Docker didn't consider the idea that you might want to
ship containers for multiple architectures until very recently and so
whichever of the current two devcontainer jobs finishes last ends up
replacing the container pushed for the other.  Just pushing two
container images with the same name and tag for different architectures
is not sufficient to tell the repository that they are the same
container image.  Instead, we need to push them with different tags and
then run some manifest commands.

Hopefully these are the right commands but the documentation is quite
sparse.