Include sonata hyperram boot stub in dev container.
The current sonata_simulator_boot_stub assumes the RTOS code is in
SRAM at 0x101000, but once we put the code in hyperram we actually
want it to jump to 0x40000000. This change includes both versions of
the boot stub in the devcontainer and symlinks the SRAM one by the old
name for backwards compatibility.